+++ CHANGELOG.rst
@@ -5,11 +5,13 @@ Next version
 ~~~~~~~~~~~~
 
 - Added the ``django-prose-editor[sanitize]`` which automatically installs the
-  ``nh3`` dependency.
+  ``nh3`` dependency. Thanks @plenaerts!
 - Properly restored the textarea element when destroying the editor.
 - Added more unittesting.
 - Supported using the ``ProseEditorFormField`` with widget instances, not just
   with widget classes.
+- Documented the CSS custom properties. Thanks @carltongibson!
+- Converted the block type dropdown back to a button group.
 
 
 0.6 (2024-07-26)